--- core/design-system/src/main/java/com/moneymong/moneymong/design_system/component/textfield/util/TextFiledDecorators.kt
@@ -1,26 +1,0 @@
-package com.moneymong.moneymong.design_system.component.textfield.util
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.withStyle
-import androidx.compose.ui.unit.dp
-import com.moneymong.moneymong.design_system.theme.Red03
-
-
-@Composable
-fun withRequiredMark(title: String) = buildAnnotatedString {
-    val density = LocalDensity.current
-    val spacingInSp = with(density) { 2.dp.toSp() }
-
-    val markColor = Red03
-
-    append(text = title)
-    withStyle(style = SpanStyle(fontSize = spacingInSp)) {
-        append("\u2003")
-    }
-    withStyle(style = SpanStyle(color = markColor)) {
-        append(text = "*")
-    }
-}
